aws-python-lambda-authorizer/requirements.txt
# PyJWT==2.1.0
# cryptography==37.0.4
python-jose